Azure-Container-Apps/requirements.txt
flask
gunicorn
openai
azure-identity
requests
